io_antarctica_scene: Major rewrite and changes in functionality
The basic workflow has not changed, but the internals have changed considerably
Highlights:
* The entire package is renamed to "SuperTuxKart Exporter Tools"
* Bump version number to 4.0; add my name into the author list
* Group all SuperTuxKart functionality and scripts as one module
* All exporters and panels are enabled and disabled at once
* Everything is prepared by an init script, more integration between all scripts
* Addition of a common utility script that can be called by other scripts here
* Materials script handles configuration and export better; no actual exporting yet
* Data XML files are now used in the materials script, removed hard-coded parameters
* Debloat and move most classes and definitions from panel script to utility script
* Common operators from kart and track scripts moved into panel script
* Kart and track exporters are disabled if their respective settings are not set
* Experimental standalone XML exporter that currently exports 'materials.xml'
* Quick exporter panel replaces the old, separated kart and track exporter panels
* Add-on preferences can now change the assets (data) directory, with file browser
* Add comment tooltips for all exporters
