+++ io_antarctica_scene/stk_material.py
@@ -33,6 +33,7 @@
 # Physical properties (related to gameplay) are stored in custom properties
 
 import bpy
+from collections import OrderedDict
 
 from bpy.props import (StringProperty,
                    BoolProperty,
@@ -46,6 +47,8 @@ from bpy.types import (Panel,
                    PropertyGroup
                    )
 
+from . import stk_utils, stk_panel
+
 bl_info = {
     "name": "SuperTuxKart Material Exporter",
     "description": "Exports material properties to the materials.xml format",
@@ -57,7 +60,6 @@ bl_info = {
     "wiki_url": "https://supertuxkart.net/Community",
     "tracker_url": "https://github.com/supertuxkart/stk-blender/issues",
     "category": "Import-Export"}
-
 
 # GUI Library
 # =============================================================================
@@ -78,7 +80,7 @@ class GuiGroup:
         new_prop["prop"] = prop
         new_prop["subprop"] = subprop
         self._properties.append(new_prop)
-    
+
     def register_all(self):
         for prop in self._properties:
             self._class.__annotations__[prop["key"]] = prop["prop"]
@@ -121,7 +123,7 @@ prop_action = EnumProperty(
                ("reset", "Rescue kart", ""),
                ("push", "Push back kart", ""),
               ],
-        default='none') 
+        default='none')
 
 
 GUI_action = GuiGroup(QueryProps)
@@ -167,7 +169,36 @@ GUI_gameplay.register_all()
 # End of GUI declaration
 # ----------------------
 
-class ANTARCTICA_PT_properties(bpy.types.Panel):
+# Detect if we are dealing with a SuperTuxKart shader
+# The custom STK PBR shader is not yet implemented
+# Use a principled BSDF shader for now
+def is_stk_shader(node):
+    if node.bl_static_type == "BSDF_PRINCIPLED":
+    #if node.bl_static_type == "GROUP":
+        # if node.node_tree.name == "stk_pbr_shader":
+        return True
+    else:
+        return False
+
+# Get the SuperTuxKart shader name
+def get_stk_shader_name(node):
+    return node.name
+
+# We make sure we get the root of the node tree (we start from the output and build up)
+def get_root_shader(node_network):
+    for node in node_network:
+        # We check if it's a material output
+        if node.bl_static_type == "OUTPUT_MATERIAL":
+            # The surface should be linked
+            if node.inputs["Surface"].is_linked:
+                # and the surface should be linked to a stk shader
+                child = node.inputs["Surface"].links[0].from_node
+                if is_stk_shader(child):
+                    return child
+
+    return None
+
+class ANTARCTICA_PT_properties(Panel, stk_panel.PanelBase):
     bl_idname = "ANTARCTICA_PT_properties"
     bl_label = "Antarctica Properties"
     bl_space_type = 'PROPERTIES'
@@ -175,9 +206,34 @@ class ANTARCTICA_PT_properties(bpy.types.Panel):
     bl_context = 'material'
 
     def draw(self, context):
-        pass
+        layout = self.layout
+        row = layout.row()
 
-class ANTARCTICA_PT_interaction_gameplay(Panel):
+        obj = stk_utils.getObject(context, stk_panel.CONTEXT_MATERIAL)
+        if obj is not None:
+            root_node = get_root_shader(obj.node_tree.nodes)
+            if root_node is not None:
+                base_color = root_node.inputs["Base Color"]
+                if base_color.is_linked:
+                    col = base_color.links[0].from_node
+                    if type(col) is bpy.types.ShaderNodeTexImage:
+                        row.label(text="Backing image: " + col.image.name)
+                    elif type(col) is bpy.types.ShaderNodeVertexColor:
+                        row.label(text="Backing image: (vertex color)")
+                    else:
+                        row.label(text="(Incompatible node detected)")
+                else:
+                    row.label(text="Backing image: (none)")
+            else:
+                row.label(text="(Material is not node-based)")
+
+            properties = OrderedDict([])
+            for curr in stk_panel.STK_MATERIAL_PROPERTIES[1]:
+                properties[curr.id] = curr
+
+            self.recursivelyAddProperties(properties, layout, obj, stk_panel.CONTEXT_MATERIAL)
+
+class ANTARCTICA_PT_interaction_gameplay(Panel, stk_panel.PanelBase):
     bl_parent_id = "ANTARCTICA_PT_properties"
     bl_idname = "ANTARCTICA_PT_interaction_gameplay"
     bl_label = "Interaction / Gameplay"
@@ -192,14 +248,14 @@ class ANTARCTICA_PT_interaction_gameplay(Panel):
     def draw(self, context):
         layout = self.layout
         layout.use_property_split = True
-        
+
         props = bpy.context.scene.QueryProps
-        
+
         box2 = layout.box()
         global GUI_gameplay
         GUI_gameplay.draw(box2)
 
-class ANTARCTICA_PT_display(Panel):
+class ANTARCTICA_PT_display(Panel, stk_panel.PanelBase):
     bl_parent_id = "ANTARCTICA_PT_properties"
     bl_idname = "ANTARCTICA_PT_display"
     bl_label = "Shader Properties"
@@ -221,6 +277,8 @@ class ANTARCTICA_PT_display(Panel):
 # Writes the materials files, which includes all texture definitions
 # Items are accessed by nodes, instead of being accessed directly
 def writeMaterialsFile(sPath):
+    used_inputs = ["Base Color", "Metallic", "Specular", "Roughness", "Emission", "Alpha", "Normal"]
+
     sp_mat_props = {
                'shader_name'            : {'default': "", 'parent': None, 'type': 'string'},
                'uv_two_tex'             : {'default': "", 'parent': None, 'type': 'string'},
@@ -228,7 +286,21 @@ def writeMaterialsFile(sPath):
                'tex_layer_3'             : {'default': "", 'parent': None, 'type': 'string'},
                'tex_layer_4'            : {'default': "", 'parent': None, 'type': 'string'},
                'tex_layer_5'             : {'default': "", 'parent': None, 'type': 'string'}
-        }
+    }
+
+    old_mat_props = {
+           'shader'                : {'default': 'solid', 'parent': None, 'type': 'string'},
+           'splatting_texture_1'   : {'default': "", 'parent': ('shader','splatting'), 'type': 'string'},
+           'splatting_texture_2'   : {'default': "", 'parent': ('shader','splatting'), 'type': 'string'},
+           'splatting_texture_3'   : {'default': "", 'parent': ('shader','splatting'), 'type': 'string'},
+           'splatting_texture_4'   : {'default': "", 'parent': ('shader','splatting'), 'type': 'string'},
+           'normal_map'            : {'default': "", 'parent': None, 'type': 'string'},
+           'gloss_map'             : {'default': "", 'parent': None, 'type': 'string'},
+           'clampu'                : {'default': "N", 'parent': None, 'type': 'bool'},
+           'clampv'                : {'default': "N", 'parent': None, 'type': 'bool'},
+           'grass_speed'           : {'default': 0.4, 'parent': ('shader','grass'), 'type': 'number'},
+           'grass_amplitude'       : {'default': 0.25, 'parent': ('shader','grass'), 'type': 'number'}
+    }
 
     other_mat_props = {
                'below_surface'         : {'default': "N", 'parent': None, 'type': 'bool'},
@@ -250,51 +322,108 @@ def writeMaterialsFile(sPath):
                'colorization_factor'   : {'default': 0.0, 'parent': 'colorizable', 'type': 'number'},
                'colorization_mask'     : {'default': "", 'parent': 'colorizable', 'type': 'string'},
                'hue_settings'          : {'default': "", 'parent': 'colorizable', 'type': 'string'}
-        }
+    }
 
     class LogReport:
 
-        def log(name):
+        def error(name):
             print("Error detected in the following material:", name)
+
+        def warn(name):
+            print("Warning reported in the following material:", name)
 
         def info(info):
             print(info)
 
         def abort(error, reason):
-            raise Exception("Material" + error, reason)
-
-    # Detect if we are dealing with a SuperTuxKart shader
-    # The custom STK PBR shader is not yet implemented
-    # Use a principled BSDF shader for now
-    def is_stk_shader(node):
-        if node.bl_static_type == "BSDF_PRINCIPLED":
-        #if node.bl_static_type == "GROUP":
-            # if node.node_tree.name == "stk_pbr_shader":
-            return True
-        else:
-            return False
-
-    # Get the SuperTuxKart shader name
-    def get_stk_shader_name(node):
-        return node.name
-
-    # We make sure we get the root of the node tree (we start from the output and build up)
-    def get_root_shader(node_network):
-        for node in node_network:
-            # We check if it's a material output
-            if node.bl_static_type == "OUTPUT_MATERIAL":
-                # The surface should be linked
-                if node.inputs["Surface"].is_linked:
-                    # and the surface should be linked to a stk shader
-                    child = node.inputs["Surface"].links[0].from_node
-                    if is_stk_shader(child):
-                        return child
-
-        return None
+            raise RuntimeError("Material" + error, reason)
 
     print("\nAntractica Material Exporter")
     print("===")
+
+    print("<?xml version=\"1.0\" encoding=\"utf-8\"?>\n")
+    print("<materials>\n")
+
     for mat in bpy.data.materials:
+        # Check if a material is using SP shader materials first
+        sp_mat = stk_utils.getIdProperty(mat, "shader", default="", set_value_if_undefined=0) == "sp_shader"
+        if sp_mat == True:
+            mat_dic = stk_utils.merge_materials(other_mat_props, sp_mat_props)
+        else:
+            mat_dic = stk_utils.merge_materials(other_mat_props, old_mat_props)
+
+        # Iterate through material definitions and collect data
+        matLine = ""
+        paramLine = ""
+        sImage = ""
+        sSFX = ""
+        sParticle = ""
+        sZipper = ""
+        hasSoundeffect = (stk_utils.convertTextToYN(stk_utils.getIdProperty(mat, "use_sfx", "no")) == "Y")
+        hasParticle = (stk_utils.convertTextToYN(stk_utils.getIdProperty(mat, "particle", "no")) == "Y")
+        hasZipper = (stk_utils.convertTextToYN(stk_utils.getIdProperty(mat, "zipper", "no")) == "Y")
+
+        # Create a copy of the list of defaults so that it can be modified. Then add
+        # all properties of the current image
+        props_copy = []
+        for sAttrib in mat.keys():
+            if sAttrib not in props_copy:
+                props_copy.append( (sAttrib, mat[sAttrib]) )
+
+        for AProperty,ADefault in props_copy:
+            # Don't add the (default) values to the property list
+            currentValue = stk_utils.getIdProperty(mat, AProperty, ADefault, set_value_if_undefined=0)
+            # Correct for all the ways booleans can be represented (true/false;yes/no;zero/not_zero)
+            if AProperty in mat_dic and mat_dic[AProperty]['type'] == 'bool':
+                currentValue = stk_utils.convertTextToYN(currentValue)
+
+            # These items pertain to the soundeffects (starting with sfx_)
+            if AProperty.strip().startswith("sfx_"):
+                strippedName = AProperty.strip()[len("sfx_"):]
+
+                if strippedName in ['filename', 'rolloff', 'min_speed', 'max_speed', 'min_pitch', 'max_pitch', 'positional', 'volume']:
+                    if isinstance(currentValue, float):
+                        sSFX = "%s %s=\"%.2f\""%(sSFX,strippedName,currentValue)
+                    else:
+                        sSFX = "%s %s=\"%s\""%(sSFX,strippedName,currentValue)
+            elif AProperty.strip().upper().startswith("PARTICLE_"):
+                #These items pertain to the particles (starting with particle_)
+                strippedName = AProperty.strip()[len("PARTICLE_"):]
+                sParticle = "%s %s=\"%s\""%(sParticle,strippedName,currentValue)
+            elif AProperty.strip().upper().startswith("ZIPPER_"):
+                #These items pertain to the zippers (starting with zipper_)
+                strippedName = AProperty.strip()[len("ZIPPER_"):]
+
+                sZipper = "%s %s=\"%s\""%(sZipper,strippedName.replace('_', '-'),currentValue)
+            else:
+                # These items are standard items
+                prop = AProperty.strip()#.lower()
+
+                if prop in mat_dic.keys():
+
+                    # If this property is conditional on another
+                    cond = mat_dic[prop]['parent']
+
+                    conditionPassed = False
+                    if cond is None:
+                        conditionPassed = True
+                    elif type(cond) is tuple:
+                        if cond[0] in mat and mat[cond[0]] == cond[1]:
+                            conditionPassed = True
+                    elif cond in mat and mat[cond] == "true":
+                        conditionPassed = True
+
+                    if currentValue != mat_dic[prop]['default'] and conditionPassed:
+                        fixed_property = AProperty
+                        if AProperty == 'shader_name':
+                            fixed_property = 'shader'
+                        if isinstance(currentValue, float):
+                            # In blender, proeprties use '_', but STK still expects '-'
+                            paramLine = "%s %s=\"%.2f\""%(paramLine,fixed_property.replace("_","-"),currentValue)
+                        else:
+                            # In blender, proeprties use '_', but STK still expects '-'
+                            paramLine = "%s %s=\"%s\""%(paramLine,fixed_property.replace("_","-"),(currentValue+'').strip())
+
         # Do not export non-node based materials
         if mat.node_tree is not None:
             root = get_root_shader(mat.node_tree.nodes)
@@ -302,94 +431,86 @@ def writeMaterialsFile(sPath):
 
             # If we can't find a root node we raise an error
             if root == None:
-                LogReport.log(mat.name)
-                LogReport.info(" Make sure you only have one 'Material Output' in your shader graph")
-                LogReport.info(" Make sure you connected a valid SuperTuxKart shader to 'Material Output'")
+                LogReport.error(mat.name)
+                LogReport.info("Make sure you only have one 'Material Output' in your shader graph")
+                LogReport.info("Make sure you connected a valid SuperTuxKart shader to 'Material Output'")
                 LogReport.abort("ShaderEditor", "We can't find a root node.")
 
-            xml_output = '<material> name="{}" shader_name="{}" '.format(mat.name, get_stk_shader_name(root))
+            #xml_output = '<material> name="{}" shader_name="{}" '.format(mat.name, get_stk_shader_name(root))
 
             count = 0 # Counter for the texture layers
             for inp in root.inputs:
                 # Only certain inputs will be used from the shader, not all of them
                 # Managing colors / 3D
-                if type(inp) is bpy.types.NodeSocketColor or type(inp) is bpy.types.NodeSocketVector:
+                if type(inp) is bpy.types.NodeSocketColor or type(inp) is bpy.types.NodeSocketVector and \
+                inp.name in used_inputs:
                     print("Export Color or Vector: " + inp.name)
                     if inp.is_linked:
                         # Get the connected node
                         child = inp.links[0].from_node
                         if type(child) is bpy.types.ShaderNodeTexImage:
-                            xml_output += 'tex-layer-{}="{}" '.format(count, child.image.name)
+                            #xml_output += 'tex-layer-{}="{}" '.format(count, child.image.name)
+                            sImage = child.image
                             count += 1
+                        elif type(child) is bpy.types.ShaderNodeMixRGB:
+                            pass
                         else:
-                            LogReport.log(mat.name)
+                            LogReport.warn(mat.name)
                             LogReport.info("Texture node not found, skipping this input node")
                     else:
-                        LogReport.log(mat.name)
+                        LogReport.warn(mat.name)
                         LogReport.info("Color or Vector input not connected, skipping")
 
                 # Managing floating point numbers
-                elif type(inp) is bpy.types.NodeSocketFloatFactor:
+                elif type(inp) is bpy.types.NodeSocketFloatFactor and inp.name in used_inputs:
                     print("Export Float: " + inp.name)
-                    xml_output += '{}="{}" '.format(inp.name, inp.default_value)
+                    #xml_output += '{}="{}" '.format(inp.name, inp.default_value)
 
+                #else:
+                #    LogReport.warn(mat.name)
+                #    LogReport.info("Unsupported node input/output type found; skipping this node")
+
+                #for k in mat.keys():
+                    #if k not in "_RNA_UI":
+                        #xml_output += '{}="{}" '.format(k, mat[k])
+
+            # Now write the main content of the materials.xml file
+            if sImage or hasSoundeffect or hasParticle or hasZipper:
+                #Get the filename of the image.
+                #iName = sImage.filepath
+                matLine = "  <material name=\"%s\"" % (sImage.name)
+                if paramLine:
+                    matLine += paramLine
+                if hasSoundeffect:
+                    matLine += "\n    <sfx%s/" % (sSFX)
+                if hasParticle:
+                    matLine += "\n    <particles%s/" % (sParticle)
+                if hasZipper:
+                    matLine += "\n    <zipper%s/" % (sZipper)
+                if not hasSoundeffect and not hasParticle and not hasZipper:
+                    matLine += "/>\n"
                 else:
-                    LogReport.log(mat.name)
-                    LogReport.info("Unsupported node input/output type found; skipping this node")
+                    matLine += "\n  </material>\n"
 
-                for k in mat.keys():
-                    if k not in "_RNA_UI":
-                        xml_output += '{}="{}" '.format(k, mat[k])
+                print(matLine)
+                #f.write(matLine)
+            #xml_output += "</material>"
+            #print(xml_output)
 
-            xml_output += "</material>"
-            print(xml_output)
         else:
-            LogReport.log(mat.name)
+            LogReport.warn(mat.name)
             LogReport.info("No node tree was found; skipping this material")
 
+    print("</materials>\n")
 
 
 class STK_Material_Export_Operator(bpy.types.Operator):
-    bl_idname = ("screen.stk_material_exporter")
+    """Export XML flies describing STK materials"""
+
+    bl_idname = ("screen.stk_material_export")
     bl_label = ("Export Materials")
     filepath: bpy.props.StringProperty()
 
     def execute(self, context):
         writeMaterialsFile(self.filepath)
         return {'FINISHED'}
-
-# Add to a menu
-def menu_func_export(self, context):
-    global the_scene
-    the_scene = context.scene
-    self.layout.operator(STK_Material_Export_Operator.bl_idname, text="STK Materials")
-
-classes = (
-    QueryProps,
-    ANTARCTICA_PT_properties,
-    ANTARCTICA_PT_display,
-    ANTARCTICA_PT_interaction_gameplay,
-    STK_Material_Export_Operator
-)
-
-def register():
-    from bpy.utils import register_class
-    for cls in classes:
-        register_class(cls)
-    # Register QueryProps
-    bpy.types.Scene.QueryProps = bpy.props.PointerProperty(type=QueryProps)
-
-    bpy.types.TOPBAR_MT_file_export.append(menu_func_export)
-
-def unregister():
-    bpy.types.TOPBAR_MT_file_export.remove(menu_func_export)
-
-    from bpy.utils import unregister_class
-    for cls in classes:
-        unregister_class(cls)
-    # $ delete QueryProps on unregister
-    del(bpy.types.Scene.QueryProps)
-
-
-if __name__ == "__main__":
-    register()
